Keep the vertices clockwise even when using negative scale.
This is in regards to: https://github.com/liabru/matter-js/issues/443

I noticed when I flipped a body over X using negative scale, the body would wobble on edges and react a bit different. After reading the above issue, I had a feeling this might be the issue so I went in and created a solution which fixed the wobble and differences.